Split the Percolator data types into their own files
* PercolatorConfig, PercolatorEntry, PercolatorResult and PercolatorResults moved
  out of PercolatorFdr.cs, which drops 4,285 -> 3,999 lines
* No call sites changed: same namespace, so nothing referencing the types moved
* Types moved verbatim; Stellar regression byte-identical on modes 1, 2 and 3

See ai/todos/active/TODO-20260727_osprey_percolatorfdr_decomposition.md
