+++ pwiz_tools/Osprey/Osprey.FDR/PercolatorFdr.cs
@@ -43,292 +43,6 @@ using pwiz.Osprey.ML;
 namespace pwiz.Osprey.FDR
 {
     /// <summary>
-    /// Configuration for the Percolator runner.
-    /// </summary>
-    public class PercolatorConfig
-    {
-        /// <summary>FDR threshold for selecting positive training set (default: 0.01).</summary>
-        public double TrainFdr { get; set; }
-
-        /// <summary>FDR threshold for final results (default: 0.01).</summary>
-        public double TestFdr { get; set; }
-
-        /// <summary>Maximum SVM training iterations per fold (default: 10).</summary>
-        public int MaxIterations { get; set; }
-
-        /// <summary>Number of cross-validation folds (default: 3).</summary>
-        public int NFolds { get; set; }
-
-        /// <summary>Random seed for reproducibility (default: 42).</summary>
-        public ulong Seed { get; set; }
-
-        /// <summary>Grid search C values for SVM cost parameter.</summary>
-        public double[] CValues { get; set; }
-
-        /// <summary>Maximum paired entries for SVM cross-validation (default: 300000).</summary>
-        public int MaxTrainSize { get; set; }
-
-        /// <summary>
-        /// Optional per-feature metadata (machine name, display label, expected
-        /// direction) in PIN-index order -- a single vector of
-        /// <see cref="OspreyFeatureInfo"/> populated by the Tasks layer from the
-        /// feature calculators (the FDR project does not reference the Scoring
-        /// assembly that owns the calculator SPI). The names drive the Stage 5
-        /// diagnostic dumps and the best-initial-feature log line; the labels +
-        /// directions drive the post-training feature-contribution report. Null
-        /// suppresses the names (logged as "unknown"/"feature_i") and the
-        /// unexpected-direction flag. Never used for scoring.
-        /// </summary>
-        public OspreyFeatureInfo[] FeatureInfos { get; set; }
-
-        /// <summary>
-        /// If true, the feature-contribution accumulator also bins each feature's
-        /// standardized value by class into per-feature target/decoy histograms
-        /// (the <c>--model-diagnostics</c> per-feature distribution view). Off for
-        /// the production scoring path so it adds no work. Never used for scoring.
-        /// </summary>
-        public bool CollectFeatureHistograms { get; set; }
-
-        /// <summary>
-        /// If true, <see cref="PercolatorFdr.RunPercolator"/> trains the fold
-        /// models + standardizer and returns early -- skips CV/averaged
-        /// scoring of the input entries, PEP estimation, q-value
-        /// computation, and per-file FDR logging. Used by the streaming
-        /// path in <c>AnalysisPipeline.RunPercolatorStreaming</c>, where
-        /// the caller pre-dedups + subsamples the entries for training,
-        /// then invokes <see cref="PercolatorFdr.ScorePopulationAndComputeFdr"/>
-        /// on the full per-file FdrEntry population. Mirrors Rust's
-        /// <c>PercolatorConfig::train_only</c>.
-        /// </summary>
-        public bool TrainOnly { get; set; }
-
-        /// <summary>
-        /// Stage 5 diagnostic-dump gates, or <c>null</c> (the common case) when
-        /// diagnostics are off. Carried in by the Tasks-layer caller so
-        /// <see cref="PercolatorFdr.RunPercolator"/> stays a pure function of its
-        /// inputs -- it reads no env vars and never exits the process. See
-        /// <see cref="PercolatorDiagnosticsConfig"/>.
-        /// </summary>
-        public PercolatorDiagnosticsConfig Diagnostics { get; set; }
-
-        /// <summary>
-        /// Train gradient-boosted decision trees (<c>--fdr-method gbdt</c>) instead
-        /// of the linear SVM. Everything else about the run is unchanged: the same
-        /// best-per-precursor dedup, the same peptide-grouped CV folds, the same
-        /// semi-supervised positive-set iteration, and the same target-decoy
-        /// competition / q-value / PEP math. Only the per-fold classifier and the
-        /// full-population scoring function differ, so the two methods are directly
-        /// comparable at matched entrapment FDP. See <see cref="GbtParams"/>.
-        /// </summary>
-        public bool UseGradientBoostedTrees { get; set; }
-
-        /// <summary>
-        /// Hyper-parameters for the <see cref="UseGradientBoostedTrees"/> classifier.
-        /// Ignored on the default (SVM) path. Never null -- defaults to the validated
-        /// conservative setting from <see cref="GbtParams"/>.
-        /// </summary>
-        public GbtParams GbtParams { get; set; }
-
-        /// <summary>
-        /// Worker threads for the gradient-boosted-trees full-population score pass
-        /// (from <c>--threads</c>). Used ONLY there: the linear score pass stays serial
-        /// because its feature-contribution accumulator is a running float sum whose
-        /// order is byte-parity-locked, whereas a tree score is a pure per-row function
-        /// that accumulates nothing. Defaults to the machine's processor count.
-        /// </summary>
-        public int NThreads { get; set; }
-
-        public PercolatorConfig()
-        {
-            TrainFdr = 0.01;
-            TestFdr = 0.01;
-            MaxIterations = 10;
-            NFolds = 3;
-            Seed = 42;
-            CValues = new[] { 0.001, 0.01, 0.1, 1.0, 10.0, 100.0 };
-            MaxTrainSize = 300000;
-            TrainOnly = false;
-            UseGradientBoostedTrees = false;
-            GbtParams = new GbtParams();
-            NThreads = Environment.ProcessorCount;
-        }
-
-        /// <summary>
-        /// A copy of this config with <see cref="TrainOnly"/> set: what the streaming
-        /// paths hand to <see cref="PercolatorFdr.RunPercolator"/> for the
-        /// train-the-fold-models pass over the subsampled training set. Copies every
-        /// knob that selects HOW a model is trained -- including the classifier choice
-        /// and its hyper-parameters -- so the training pass cannot silently diverge from
-        /// the scoring pass that consumes its output. (Both streaming paths previously
-        /// hand-copied this field list, which meant a new training knob had to be added
-        /// in two places or one path would quietly train the wrong model.)
-        ///
-        /// <see cref="CollectFeatureHistograms"/> is deliberately NOT carried: the
-        /// histograms are accumulated by the score pass over the full population, not
-        /// over the training subset.
-        /// </summary>
-        public PercolatorConfig CloneForTrainOnly()
-        {
-            return new PercolatorConfig
-            {
-                TrainFdr = TrainFdr,
-                TestFdr = TestFdr,
-                MaxIterations = MaxIterations,
-                NFolds = NFolds,
-                Seed = Seed,
-                CValues = CValues,
-                MaxTrainSize = MaxTrainSize,
-                FeatureInfos = FeatureInfos,
-                UseGradientBoostedTrees = UseGradientBoostedTrees,
-                GbtParams = GbtParams,
-                NThreads = NThreads,
-                TrainOnly = true,
-                Diagnostics = Diagnostics
-            };
-        }
-    }
-
-    /// <summary>
-    /// Input entry for Percolator scoring.
-    /// </summary>
-    public class PercolatorEntry
-    {
-        /// <summary>Source file name (for per-run FDR).</summary>
-        public string FileName { get; set; }
-
-        /// <summary>Modified peptide sequence (for fold grouping and peptide-level FDR).</summary>
-        public string Peptide { get; set; }
-
-        /// <summary>Precursor charge state.</summary>
-        public byte Charge { get; set; }
-
-        /// <summary>Whether this is a decoy.</summary>
-        public bool IsDecoy { get; set; }
-
-        /// <summary>Entry ID for target-decoy pairing (high bit = decoy).</summary>
-        public uint EntryId { get; set; }
-
-        /// <summary>
-        /// Row index of this observation within its source file's
-        /// <c>.scores.parquet</c>. Lets the streaming score pass reload the
-        /// 21-feature vector on demand (issue #4355 Phase 4) instead of holding
-        /// every entry's <see cref="Features"/> resident for the whole join.
-        /// <c>uint.MaxValue</c> marks an appended entry (e.g. Stage 6 gap-fill)
-        /// that has no original parquet row.
-        /// </summary>
-        public uint ParquetIndex { get; set; }
-
-        /// <summary>
-        /// The <c>fragment_coelution_sum</c> feature (PIN feature 0), carried as a
-        /// resident scalar so best-per-precursor selection can rank observations
-        /// without holding the full <see cref="Features"/> vector. Equal to
-        /// <c>Features[0]</c> byte-for-byte on the first pass (both come from the
-        /// same parquet column / <c>CoelutionScorer</c> assignment).
-        /// </summary>
-        public double CoelutionSum { get; set; }
-
-        /// <summary>Raw feature values. Null on streaming-path stubs (issue #4355
-        /// Phase 4), where the vector is loaded per file from parquet at score
-        /// time by <see cref="ParquetIndex"/>.</summary>
-        public double[] Features { get; set; }
-    }
-
-    /// <summary>
-    /// Result for a single entry after Percolator scoring.
-    /// </summary>
-    public class PercolatorResult
-    {
-        /// <summary>SVM decision function score.</summary>
-        public double Score { get; set; }
-
-        /// <summary>Per-run precursor-level q-value.</summary>
-        public double RunPrecursorQvalue { get; set; }
-
-        /// <summary>Per-run peptide-level q-value.</summary>
-        public double RunPeptideQvalue { get; set; }
-
-        /// <summary>Experiment-wide precursor-level q-value.</summary>
-        public double ExperimentPrecursorQvalue { get; set; }
-
-        /// <summary>Experiment-wide peptide-level q-value.</summary>
-        public double ExperimentPeptideQvalue { get; set; }
-
-        /// <summary>Posterior error probability.</summary>
-        public double Pep { get; set; }
-
-        public PercolatorResult()
-        {
-            RunPrecursorQvalue = 1.0;
-            RunPeptideQvalue = 1.0;
-            ExperimentPrecursorQvalue = 1.0;
-            ExperimentPeptideQvalue = 1.0;
-            Pep = 1.0;
-        }
-    }
-
-    /// <summary>
-    /// Full results from Percolator analysis.
-    /// </summary>
-    public class PercolatorResults
-    {
-        /// <summary>Per-entry results.</summary>
-        public List<PercolatorResult> Entries { get; set; }
-
-        /// <summary>Feature weights from best model per fold. Empty on the
-        /// gradient-boosted-trees path, which has no linear weights -- see
-        /// <see cref="FoldGbtModels"/>.</summary>
-        public List<double[]> FoldWeights { get; set; }
-
-        /// <summary>Bias terms from best model per fold. Empty on the
-        /// gradient-boosted-trees path.</summary>
-        public List<double> FoldBiases { get; set; }
-
-        /// <summary>
-        /// Best gradient-boosted-trees model per fold when the run was configured with
-        /// <see cref="PercolatorConfig.UseGradientBoostedTrees"/>; <c>null</c> on the
-        /// default SVM path. The two are mutually exclusive: whichever is populated
-        /// selects how the score passes turn a standardized feature row into a score.
-        ///
-        /// The linear path averages the fold WEIGHTS and scores once; trees cannot be
-        /// averaged that way, so the tree path averages the fold SCORES instead. For a
-        /// linear model those are the same operation (a dot product is linear in the
-        /// weights), which is why the SVM path keeps its cheaper weight-average form.
-        /// </summary>
-        public List<GradientBoostedTrees> FoldGbtModels { get; set; }
-
-        /// <summary>
-        /// The Skyline-style per-feature percent-contribution decomposition of the
-        /// trained averaged model, or <c>null</c> when scoring did not run (e.g. the
-        /// <c>TrainOnly</c> early-return or the empty-population shortcut). Carries
-        /// the decomposition object the report is printed from; not used by scoring.
-        /// </summary>
-        public FeatureContributions FeatureContributions { get; set; }
-
-        /// <summary>Feature standardizer used during training.</summary>
-        public FeatureStandardizer Standardizer { get; set; }
-
-        /// <summary>Number of iterations used per fold.</summary>
-        public List<int> IterationsPerFold { get; set; }
-
-        /// <summary>
-        /// Set when <see cref="PercolatorFdr.RunPercolator"/> wrote a
-        /// diagnostic-only (<c>*Only</c>) dump and stopped early instead of
-        /// completing FDR. The Tasks-layer caller inspects this and performs the
-        /// process early-exit; the engine itself never exits. The other fields are
-        /// unpopulated when this is <c>true</c>.
-        /// </summary>
-        public bool DiagnosticAbort { get; set; }
-
-        public PercolatorResults()
-        {
-            Entries = new List<PercolatorResult>();
-            FoldWeights = new List<double[]>();
-            FoldBiases = new List<double>();
-            IterationsPerFold = new List<int>();
-        }
-    }
-
-    /// <summary>
     /// Performs false discovery rate estimation using the Percolator algorithm.
     /// Port of osprey-fdr/src/percolator.rs.
     /// </summary>
